refactor optimize Redis caching using priority based TTL per news source

+++ news-fetcher-service/src/main/java/com/newshunter/news_fetcher_service/entity/RssSource.java
@@ -18,14 +18,14 @@ public class RssSource {
     private final int rateLimitWindowMinutes;   // time window in minutes
 
     // TTL per source (in hours) to delete data from cache
-    private final int ttlHOURS;
+    private final long ttlHOURS;
 
 
     /**
      * Sets lastFetchedAtMinutes to current time to prevent immediate fetch on app start
      * this so important
      */
-    public RssSource(String url, int fetchIntervalMinutes, int maxRequests, int rateLimitWindowMinutes , int ttlHOURS) {
+    public RssSource(String url, int fetchIntervalMinutes, int maxRequests, int rateLimitWindowMinutes , long ttlHOURS) {
         this.url = url;
         this.fetchIntervalMinutes = fetchIntervalMinutes;
         this.lastFetchedAtMinutes = currentMinutes();
